M1 GL
Quelques modifications dans la correction de l'exercice sur le pattern observer

+++ M1/GenieLogiciel/Corrections/DesignPatterns/Observer/PatronObservateurCorrige/src/conversionDevisesAvecAPI/Euro.java
@@ -2,12 +2,6 @@ package conversionDevisesAvecAPI;
 
 import java.util.Observable;
 
-/**
- * L'Observé est un Euro. La valeur que l'on considère est une valeur "double"
- * que l'on affichera dans la devise choisie.
- *
- * @author Yvan
- */
 public class Euro extends Observable {
 
   private double valeur;
@@ -20,17 +14,13 @@ public class Euro extends Observable {
     this.valeur = valeur;
   }
 
-  /**
-   * La méthode redéfinie transforme le type de retour que l'on sait etre un
-   * Double. Elle est également rendue publique.
-   */
   public Double getValeur() {
     return valeur;
   }
 
   public void setValeur(double valeur) {
     if (valeur != this.valeur) {
-      this.valeur = (Double) valeur;
+      this.valeur = valeur;
       setChanged();
       notifyObservers();
     }
